[ADD] impl metodo para mantener la sesion
