+++ workflix/app/src/main/java/tec/ispc/workflix/views/ui/login/LoginActivity.java
@@ -1,8 +1,6 @@
 package tec.ispc.workflix.views.ui.login;
 
-import android.content.Context;
 import android.content.Intent;
-import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
@@ -235,12 +233,6 @@ public class LoginActivity extends AppCompatActivity {
         Intent catalogoIntent = new Intent(this, CatalogoActivity.class);
         startActivity(catalogoIntent);
     }
-    public void guardarEstadoBton(){
-        SharedPreferences preferences = getSharedPreferences("sesion", Context.MODE_PRIVATE);
-        boolean estado = true;
-        SharedPreferences.Editor editor = preferences.edit();
-        editor.putBoolean("estado_usu", estado);
-        editor.commit();
-    }
+
 
 }
